[2/2] Rename ServiceAppController to ServiceController
Rename module
