+++ .mypy.ini
@@ -47,7 +47,7 @@ modules =
     azul.time,
     azul.uuids,
     azul.plugins,
-    azul.service.app_controller,
+    azul.service.controller,
     azul.digests,
     lambdas.indexer.app,
     azul.schemas,
